+++ CHANGES.md
@@ -1,3 +1,7 @@
+## PENDING
+
+* Add support for CloudFormation YAML extensions (e.g. `!Ref`).
+
 ## 0.9.5 (2016-09-26)
 
 * Add `--with-role` option, to assume a role for stack operations.